+++ src/components/Admin/index.ts
@@ -4,4 +4,5 @@ export { default as TentTypeForm } from './TentTypeForm';
 export { default as ExtrasForm } from './ExtrasForm';
 export { default as TentsManager } from './TentsManager';
 export { default as ExtrasManager } from './ExtrasManager';
-export { default as TentTypesManager } from './TentTypesManager'; 
+export { default as TentTypesManager } from './TentTypesManager';
+export { default as EnquiriesManager } from './EnquiriesManager';